+++ modules/apps/knowledge-base/knowledge-base-item-selector-api/src/main/java/com/liferay/knowledge/base/item/selector/KBAttachmentItemSelectorCriterion.java
